Validation,training,Accuracy
The dataset is divided into 80% training, 10% validation, and 10% testing sets to ensure effective model development and evaluation. The training set is used to teach the model and optimize its parameters, enabling it to learn patterns from the data. The validation set helps fine-tune hyperparameters and monitor the model’s performance during training, while the testing set assesses the final accuracy and generalization capability of the model on unseen dat